Feature/integration tests

+++ src/main/java/com/mf/minutefictionbackend/services/ThemeService.java
@@ -3,9 +3,12 @@ package com.mf.minutefictionbackend.services;
 import com.mf.minutefictionbackend.dtos.inputDtos.ThemeInputDto;
 import com.mf.minutefictionbackend.dtos.mappers.ThemeMapper;
 import com.mf.minutefictionbackend.dtos.outputDtos.ThemeOutputDto;
+import com.mf.minutefictionbackend.exceptions.BadRequestException;
 import com.mf.minutefictionbackend.exceptions.ResourceNotFoundException;
+import com.mf.minutefictionbackend.models.Story;
 import com.mf.minutefictionbackend.models.Theme;
 import com.mf.minutefictionbackend.repositories.ThemeRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -25,6 +28,7 @@ public class ThemeService {
     }
 
 
+    @Transactional
     public ThemeOutputDto createTheme(ThemeInputDto themeInputDto) {
         Theme theme = themeRepository.save(themeFromInputDtoToModel(themeInputDto));
         return ThemeMapper.themeFromModelToOutputDto(theme);
@@ -35,19 +39,24 @@ public class ThemeService {
         return ThemeMapper.themeModelListToOutputList(allThemes);
     }
 
-    public ThemeOutputDto getThemeById(Long id) {
-        Theme theme = themeRepository.findById(id)
-                        .orElseThrow(() -> new ResourceNotFoundException("No theme found with id " + id));
+    public ThemeOutputDto getThemeById(Long themeId) {
+        Theme theme = themeRepository.findById(themeId)
+                        .orElseThrow(() -> new ResourceNotFoundException("No theme found with id " + themeId));
         return ThemeMapper.themeFromModelToOutputDto(theme);
     }
 
-
+    @Transactional
     public void deleteThemeById(Long themeId) {
         Theme theme = themeRepository.findById(themeId)
                 .orElseThrow(() -> new ResourceNotFoundException("No theme found with id " + themeId));
+        List<Story> stories = theme.getStories();
+        if(stories != null && !stories.isEmpty()) {
+            throw new BadRequestException("Theme has stories added to it. Delete these first.");
+        }
         themeRepository.delete(theme);
     }
 
+    @Transactional
     public ThemeOutputDto updateTheme(Long themeId, ThemeInputDto updatedThemeInputDto) {
         Theme updateTheme = themeRepository.findById(themeId)
                 .orElseThrow(() -> new ResourceNotFoundException("No theme found"));
